requirements.txt
aenum==3.1.15
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
idna==3.6
imageio==2.33.0
lazy_loader==0.3
marshmallow==3.20.1
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pyproj==3.6.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
scikit-image==0.22.0
scipy==1.11.4
sentinelhub==3.9.4
shapely==2.0.2
six==1.16.0
tifffile==2023.9.26
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
utm==0.7.0
